Extend MainActivity and GameActivity from BaseActivity class
